// angular/src/app/main/supplyChain/sales/shared/dtos/saleAccounts-dto.ts
export class saleAccountsDto {
    id: number | undefined;
    locId: number | undefined;
    locDesc: string | undefined;
    typeId: string | undefined;
    typeDesc: string | undefined;
    salesACC: string | undefined;
    salesACCDesc: string | undefined;
    salesRetACC: string | undefined;
    salesRetACCDesc: string | undefined;
    cogsacc: string | undefined;
    cogsaccDesc: string | undefined;
    chAccountID: string | undefined;
    chAccountIDDesc: string | undefined;
    discAcc: string | undefined;
    discAccDesc: string | undefined;
    writeOffAcc: string | undefined;
    writeOffAccDesc: string | undefined;
    payableAcc: string | undefined;
    payableAccDesc: string | undefined;
    refundableAcc:string | undefined;
    refundableAccDesc:string | undefined;
}